test/web-platform-tests/to-upstream/css/cssom-view/Element-clientLeft.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Element.prototype.clientLeft</title>
<link rel=help href="https://drafts.csswg.org/cssom-view/#dom-element-clientleft">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
"use strict";

test(() => {

  const el = document.createElement("p");
  assert_equals(el.clientLeft, 0);

}, "clientLeft should return 0 for an element with no associated CSS layout box");
</script>
